tools: refactor policy generator
